finish our-work

+++ lib/delete-file.php
@@ -1,7 +1,30 @@
 <?php
 
-if (unlink($_POST['fileName'])) {
-  echo json_encode('Delete successfully');
-} else {
-  echo json_encode('Failed to delete file.');
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+  $success = [];
+  $failure = [];
+
+  if (isset($_POST['fileNames'])) {
+
+    foreach (json_decode($_POST['fileNames']) as $fileName) {
+      if (unlink($fileName)) {
+        $success[] = $fileName;
+      } else {
+        $failure[] = $fileName;
+      }
+    }
+    
+  } else {
+    
+    if (unlink($_POST['fileName'])) {
+      $success[] = $_POST['fileName'];
+    } else {
+      $failure[] = $_POST['fileName'];
+    }
+    
+  }
+  
+  echo json_encode(['success' => $success, 'failure' => $failure]);
 }
+
+exit();
